Added states animation. MA-590
