+++ css/main.css
@@ -299,7 +299,7 @@ section.hero-section {
       -ms-transform: translate(-50%, -50%);
       -o-transform: translate(-50%, -50%);
       transform: translate(-50%, -50%); }
-      .us-map-section .map-holder .states img {
+      .us-map-section .map-holder .states .map-img {
         width: 100%; }
       .us-map-section .map-holder .states .map {
         position: absolute;
@@ -1073,15 +1073,15 @@ section.full-width {
   transform: translateZ(0);
 }
 
-.state_1 {-webkit-animation-delay: 2s;animation-delay: 2s;}
-.state_2 {-webkit-animation-delay: 5s;animation-delay: 5s;}
-.state_3 {-webkit-animation-delay: 9s;animation-delay: 9s;}
-.state_4 {-webkit-animation-delay: 3s;animation-delay: 3s;}
-.state_5 {-webkit-animation-delay: 5.3s;animation-delay: 5.3s;}
-.state_6 {-webkit-animation-delay: 4s;animation-delay: 4s;}
-.state_7 {-webkit-animation-delay: 2.5s;animation-delay: 2.5s;}
-.state_8 {-webkit-animation-delay: 6s;animation-delay: 6s;}
-.state_9 {-webkit-animation-delay: 4.5s;animation-delay: 4.5s;}
+.state_1 {-webkit-animation-delay: 2s;animation-delay: 2s; width: 12%; height: 30%; left: 7%; top: 30%;}
+.state_2 {-webkit-animation-delay: 5s;animation-delay: 5s; width: 12%; height: 22%; left: 25.7%; top: 50%;}
+.state_3 {-webkit-animation-delay: 9s;animation-delay: 9s; width: 13%; height: 17%; left: 27%; top: 34.3%;}
+.state_4 {-webkit-animation-delay: 3s;animation-delay: 3s; width: 11%; height: 12%; left: 50%; top: 28%;}
+.state_5 {-webkit-animation-delay: 5.3s;animation-delay: 5.3s; width: 8%; height: 15%; left: 71.6%; top: 27.9%;}
+.state_6 {-webkit-animation-delay: 4s;animation-delay: 4s; width: 11%; height: 12%; left: 78.95%; top: 24.5%;}
+.state_7 {-webkit-animation-delay: 2.5s;animation-delay: 2.5s; width: 14%; height: 14%; left: 75%; top: 38%;}
+.state_8 {-webkit-animation-delay: 6s;animation-delay: 6s; width: 16%; height: 12%; left: 75%; top: 47%;}
+.state_9 {-webkit-animation-delay: 4.5s;animation-delay: 4.5s; width: 17%; height: 23%; left: 69%; top: 74%;}
 
 
 @-webkit-keyframes stateAnimation {
