+++ index.html
@@ -62,38 +62,11 @@
 
 		<div class="map-holder">
 			<div class="states">
-				<!-- <div class="map-states">
-					<img src="assets/img/states/state1.png" class="state state_1">
-					<img src="assets/img/states/state2.png" class="state state_2">
-					<img src="assets/img/states/state3.png" class="state state_3">
-					<img src="assets/img/states/state4.png" class="state state_4">
-					<img src="assets/img/states/state5.png" class="state state_5">
-					<img src="assets/img/states/state6.png" class="state state_6">
-					<img src="assets/img/states/state7.png" class="state state_7">
-					<img src="assets/img/states/state8.png" class="state state_8">
-					<img src="assets/img/states/state9.png" class="state state_9">
-				</div> -->
 				<div class="map">
-					<img src="assets/img/states/map.png">
+					<img src="assets/img/states/map.png" class="map-img">
 					<canvas id="map" style="width: 100%; position: absolute; left: 0; top: 0;" width="2422" height="1474"></canvas>
 				</div>
 			</div>
-			<!-- <div class="states">
-				<div class="map js-candidates">
-					<img src="assets/img/states/map.png">
-				</div>
-				<div class="map-states">
-					<img src="assets/img/states/state1.png" class="state state_1">
-					<img src="assets/img/states/state2.png" class="state state_2">
-					<img src="assets/img/states/state3.png" class="state state_3">
-					<img src="assets/img/states/state4.png" class="state state_4">
-					<img src="assets/img/states/state5.png" class="state state_5">
-					<img src="assets/img/states/state6.png" class="state state_6">
-					<img src="assets/img/states/state7.png" class="state state_7">
-					<img src="assets/img/states/state8.png" class="state state_8">
-					<img src="assets/img/states/state9.png" class="state state_9">
-				</div>
-			</div> -->
 			<div class="post-block">
 				<div class="center-block text-center">
 					<h2>Digital Video Will Decide the Next President</h2>
@@ -839,6 +812,7 @@
 	<script src="//ajax.googleapis.com/ajax/libs/jqueryui/1.11.4/jquery-ui.min.js"></script>
 	<script src="scripts/full/highcharts.js"></script>
 	<script src="scripts/full/main.js"></script>
+	<script src="scripts/full/map.js"></script>
 
 	<!-- Scripts -->
 </body>
